feat: rename module to universal_resolver
